dspy-backend/pyproject.toml
[project]
name = "asmbli-dspy-backend"
version = "0.1.0"
description = "DSPy-powered AI backend for Asmbli"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dspy-ai>=2.5.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "chromadb>=0.4.22",
    "openai>=1.12.0",
    "anthropic>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
